feat: creating custom server manager and move functions and update references
