+++ client/service/service.go
@@ -13,7 +13,6 @@ import (
 	"errors"
 	"fmt"
 	"log/slog"
-	"os"
 	"path/filepath"
 	"runtime"
 	"slices"
@@ -22,11 +21,9 @@ import (
 
 	box "github.com/sagernet/sing-box"
 	"github.com/sagernet/sing-box/adapter"
-	"github.com/sagernet/sing-box/constant"
 	"github.com/sagernet/sing-box/experimental/libbox"
 	"github.com/sagernet/sing-box/log"
 	"github.com/sagernet/sing-box/option"
-	"github.com/sagernet/sing-box/protocol/group"
 	"github.com/sagernet/sing/common/json"
 	"github.com/sagernet/sing/service"
 	"github.com/sagernet/sing/service/pause"
@@ -49,10 +46,6 @@ type BoxService struct {
 	pauseTimer        *time.Timer
 	mu                sync.Mutex
 	isRunning         bool
-
-	customServersMutex    sync.Locker
-	customServers         map[string]option.Options
-	customServersFilePath string
 }
 
 const CustomSelectorTag = "custom_selector"
@@ -70,12 +63,9 @@ type CustomServerInfo struct {
 // to interact with the underlying platform
 func New(config, dataDir string, platIfce libbox.PlatformInterface, rulesetManager *ruleset.Manager) (*BoxService, error) {
 	bs := &BoxService{
-		config:                config,
-		platIfce:              platIfce,
-		mutRuleSetManager:     rulesetManager,
-		customServersMutex:    new(sync.Mutex),
-		customServers:         make(map[string]option.Options),
-		customServersFilePath: filepath.Join(dataDir, "data", "custom_servers.json"),
+		config:            config,
+		platIfce:          platIfce,
+		mutRuleSetManager: rulesetManager,
 	}
 
 	setupOpts := &libbox.SetupOptions{
@@ -202,229 +192,6 @@ func (bs *BoxService) Wake() {
 		bs.pauseTimer.Stop()
 		bs.pauseTimer = nil
 	}
-}
-
-// ServerConnectConfig represents configuration for connecting to a custom server.
-type ServerConnectConfig []byte
-
-// AddCustomServer load or parse the given configuration and add given
-// endpdoint/outbound to the instance. We're only expecting one endpoint or
-// outbound per call.
-func (bs *BoxService) AddCustomServer(tag string, cfg ServerConnectConfig) error {
-	bs.customServersMutex.Lock()
-	loadedOptions, configExist := bs.customServers[tag]
-	bs.customServersMutex.Unlock()
-	if configExist && cfg != nil {
-		if err := bs.RemoveCustomServer(tag); err != nil {
-			return err
-		}
-	}
-
-	if cfg != nil {
-		var err error
-		loadedOptions, err = json.UnmarshalExtendedContext[option.Options](bs.ctx, cfg)
-		if err != nil {
-			return fmt.Errorf("failed to unmarshal config: %w", err)
-		}
-	}
-
-	if err := updateOutboundsEndpoints(bs.ctx, loadedOptions.Outbounds, loadedOptions.Endpoints); err != nil {
-		return fmt.Errorf("failed to update outbounds/endpoints: %w", err)
-	}
-
-	bs.customServersMutex.Lock()
-	bs.customServers[tag] = loadedOptions
-	bs.customServersMutex.Unlock()
-	if err := bs.storeCustomServer(tag, loadedOptions); err != nil {
-		return fmt.Errorf("failed to store custom server: %w", err)
-	}
-
-	return nil
-}
-
-func (bs *BoxService) ListCustomServers() ([]CustomServerInfo, error) {
-	loadedServers, err := bs.loadCustomServer()
-	if err != nil {
-		return nil, fmt.Errorf("failed to load custom servers: %w", err)
-	}
-
-	return loadedServers.CustomServers, nil
-}
-
-// storeCustomServer stores the custom server configuration to a JSON file.
-func (bs *BoxService) storeCustomServer(tag string, options option.Options) error {
-	servers, err := bs.loadCustomServer()
-	if err != nil {
-		return fmt.Errorf("load custom servers: %w", err)
-	}
-
-	if len(servers.CustomServers) == 0 {
-		servers.CustomServers = make([]CustomServerInfo, 0)
-		servers.CustomServers = append(servers.CustomServers, CustomServerInfo{
-			Tag:     tag,
-			Options: options,
-		})
-	} else {
-		for i, server := range servers.CustomServers {
-			if server.Tag == tag {
-				server.Options = options
-				servers.CustomServers[i] = server
-				break
-			}
-		}
-	}
-
-	if err = bs.writeChanges(servers); err != nil {
-		return fmt.Errorf("failed to add custom server %q: %w", tag, err)
-	}
-
-	return nil
-}
-
-func (bs *BoxService) writeChanges(customServers customServers) error {
-	storedCustomServers, err := json.MarshalContext(bs.ctx, customServers)
-	if err != nil {
-		return fmt.Errorf("marshal custom servers: %w", err)
-	}
-	if err := os.WriteFile(bs.customServersFilePath, storedCustomServers, 0644); err != nil {
-		return fmt.Errorf("write custom servers file: %w", err)
-	}
-	return nil
-}
-
-// loadCustomServer loads the custom server configuration from a JSON file.
-func (bs *BoxService) loadCustomServer() (customServers, error) {
-	var cs customServers
-	// read file and generate []byte
-	storedCustomServers, err := os.ReadFile(bs.customServersFilePath)
-	if err != nil {
-		if errors.Is(err, os.ErrNotExist) {
-			// file not exist, return empty custom servers
-			return cs, nil
-		}
-		return cs, fmt.Errorf("read custom servers file: %w", err)
-	}
-
-	if err := json.UnmarshalContext(bs.ctx, storedCustomServers, &cs); err != nil {
-		return cs, fmt.Errorf("decode custom servers file: %w", err)
-	}
-
-	bs.customServersMutex.Lock()
-	defer bs.customServersMutex.Unlock()
-	for _, v := range cs.CustomServers {
-		bs.customServers[v.Tag] = v.Options
-	}
-
-	return cs, nil
-}
-
-func (bs *BoxService) removeCustomServer(tag string) error {
-	customServers, err := bs.loadCustomServer()
-	if err != nil {
-		return fmt.Errorf("load custom servers: %w", err)
-	}
-	for i, server := range customServers.CustomServers {
-		if server.Tag == tag {
-			customServers.CustomServers = append(customServers.CustomServers[:i], customServers.CustomServers[i+1:]...)
-			break
-		}
-	}
-	if err = bs.writeChanges(customServers); err != nil {
-		return fmt.Errorf("failed to write custom server %q removal: %w", tag, err)
-	}
-	return nil
-}
-
-// RemoveCustomServer removes the custom server options from endpoints, outbounds
-// and the custom server file.
-func (bs *BoxService) RemoveCustomServer(tag string) error {
-	outboundManager := service.FromContext[adapter.OutboundManager](bs.ctx)
-	endpointManager := service.FromContext[adapter.EndpointManager](bs.ctx)
-
-	bs.customServersMutex.Lock()
-	options := bs.customServers[tag]
-	bs.customServersMutex.Unlock()
-	// selector must be removed in order to remove dependent outbounds
-	if err := outboundManager.Remove(CustomSelectorTag); err != nil {
-		return fmt.Errorf("failed to remove selector outbound: %w", err)
-	}
-	for _, outbounds := range options.Outbounds {
-		if err := outboundManager.Remove(outbounds.Tag); err != nil && !errors.Is(err, os.ErrInvalid) {
-			return fmt.Errorf("failed to remove %q outbound: %w", tag, err)
-		}
-	}
-
-	for _, endpoints := range options.Endpoints {
-		if err := endpointManager.Remove(endpoints.Tag); err != nil && !errors.Is(err, os.ErrInvalid) {
-			return fmt.Errorf("failed to remove %q endpoint: %w", tag, err)
-		}
-	}
-
-	bs.customServersMutex.Lock()
-	delete(bs.customServers, tag)
-	bs.customServersMutex.Unlock()
-	if err := bs.removeCustomServer(tag); err != nil {
-		return fmt.Errorf("failed to remove custom server %q: %w", tag, err)
-	}
-	return nil
-}
-
-// SelectCustomServer update the selector outbound to use the selected
-// outbound based on provided tag. A selector outbound must exist before
-// calling this function, otherwise it'll return a error.
-func (bs *BoxService) SelectCustomServer(tag string) error {
-	outboundManager := service.FromContext[adapter.OutboundManager](bs.ctx)
-	outbounds := outboundManager.Outbounds()
-	tags := make([]string, 0)
-	for _, outbound := range outbounds {
-		// ignoring selector because it'll be removed and re-added with the new tags
-		if outbound.Tag() == CustomSelectorTag {
-			continue
-		}
-		tags = append(tags, outbound.Tag())
-	}
-
-	if _, exists := outboundManager.Outbound(CustomSelectorTag); exists {
-		if err := outboundManager.Remove(CustomSelectorTag); err != nil {
-			return fmt.Errorf("failed to remove selector outbound: %w", err)
-		}
-	}
-
-	err := bs.newSelectorOutbound(outboundManager, CustomSelectorTag, &option.SelectorOutboundOptions{
-		Outbounds:                 tags,
-		Default:                   tag,
-		InterruptExistConnections: true,
-	})
-	if err != nil {
-		return fmt.Errorf("failed to create selector outbound: %w", err)
-	}
-
-	outbound, ok := outboundManager.Outbound(CustomSelectorTag)
-	if !ok {
-		return fmt.Errorf("failed to get selector outbound: %w", err)
-	}
-	selector, ok := outbound.(*group.Selector)
-	if !ok {
-		return fmt.Errorf("expected outbound of type *group.Selector: %w", err)
-	}
-	if err = selector.Start(); err != nil {
-		return fmt.Errorf("failed to start selector outbound: %w", err)
-	}
-	if ok = selector.SelectOutbound(tag); !ok {
-		return fmt.Errorf("failed to select outbound %q: %w", tag, err)
-	}
-
-	return nil
-}
-
-func (bs *BoxService) newSelectorOutbound(outboundManager adapter.OutboundManager, tag string, options *option.SelectorOutboundOptions) error {
-	router := service.FromContext[adapter.Router](bs.ctx)
-	logFactory := service.FromContext[log.Factory](bs.ctx)
-	if err := outboundManager.Create(bs.ctx, router, logFactory.NewLogger(tag), tag, constant.TypeSelector, options); err != nil {
-		return fmt.Errorf("create selector outbound: %w", err)
-	}
-
-	return nil
 }
 
 func (bs *BoxService) Ctx() context.Context {
